chore: work on playground class and misc
