+++ testing/cypress/pageObjects/playgroundPage.ts
@@ -1,3 +1,5 @@
+import { authenticator } from 'otplib';
+
 class PlaygroundPage {
   path: string = 'https://bcgov.github.io/keycloak-example-apps/';
 
@@ -10,13 +12,13 @@ class PlaygroundPage {
     cy.get('div').contains('Keycloak OIDC Config').click({ force: true });
   }
 
-  setAuthServerUrl(url: string) {
+  setAuthServerUrl(url: string = 'https://dev.sandbox.loginproxy.gov.bc.ca/auth') {
     cy.get(this.authServerUrl)
       .clear()
       .type(url + '{enter}');
   }
 
-  setRealm(realm: string) {
+  setRealm(realm: string = 'standard') {
     cy.get(this.realm)
       .clear()
       .type(realm + '{enter}');
@@ -30,14 +32,17 @@ class PlaygroundPage {
 
   clickUpdate() {
     cy.contains(this.commonButton, 'Update', { timeout: 10000 }).click();
+    cy.wait(2000); // Wait a bit because to make sure the data is loaded
   }
 
   clickLogin() {
     cy.contains(this.commonButton, 'Login', { timeout: 10000 }).click();
+    cy.wait(2000); // Wait a bit because to make sure the page is loaded
   }
 
   clickLogout() {
     cy.contains(this.commonButton, 'Logout', { timeout: 10000 }).click();
+    cy.wait(2000);
   }
 
   checkProviders(providers: string[]) {
@@ -94,6 +99,24 @@ class PlaygroundPage {
     cy.get('#app_totp', { timeout: 10000 }).type(token, { log: false });
     cy.contains('Verify').click();
   }
+
+  fillInPlayground = (url, realm, client) => {
+    this.selectConfig();
+    this.setAuthServerUrl(url);
+    this.setRealm(realm);
+    this.setClientId(client);
+    this.clickUpdate();
+  };
+
+  loginGithub = (username: string, password: string, secret: string) => {
+    cy.get('input#login_field').type(username, { log: false });
+    cy.get('input#password').type(password, { log: false });
+    cy.get('input[type="submit"]').click();
+
+    const token = authenticator.generate(secret);
+    cy.get('#app_totp').type(token);
+    cy.contains('Verify').click();
+  };
 }
 
 export default PlaygroundPage;
