+++ testing/cypress/support/commands.ts
@@ -27,7 +27,7 @@ Cypress.Commands.add('login', (username, password, host, siteminder) => {
   // Validate the login proxy only when we are not targeting a local install
   if (Cypress.env('host') != 'http://localhost:3000') {
     cy.origin(Cypress.env('loginproxy'), () => {
-      cy.get('#kc-header-wrapper').contains('Common Hosted Single Sign-on').should('be.visible');
+      cy.get('#kc-header-wrapper').contains('COMMON HOSTED SINGLE SIGN-ON').should('be.visible');
       cy.get('#social-idir').click();
     });
   }
